refactor: split config (validate.go) + bgp/wire (types/encode/decode)
